feat: composition errors are printed in the session that caused the error
