+++ src/command/dev/introspect.rs
@@ -1,7 +1,7 @@
 use reqwest::blocking::Client;
 use saucer::anyhow;
 
-use crate::command::dev::socket::{SubgraphSdl, SubgraphUrl};
+use crate::command::dev::protocol::{SubgraphSdl, SubgraphUrl};
 use crate::command::graph::Introspect as GraphIntrospect;
 use crate::command::subgraph::Introspect as SubgraphIntrospect;
 use crate::options::IntrospectOpts;
