--- src/command/dev/socket.rs
@@ -1,431 +1,0 @@
-use crate::{
-    command::dev::{compose::ComposeRunner, do_dev::log_err_and_continue},
-    error::RoverError,
-    Result,
-};
-use apollo_federation_types::{
-    build::SubgraphDefinition,
-    config::{FederationVersion, SupergraphConfig},
-};
-use interprocess::local_socket::{LocalSocketListener, LocalSocketStream};
-use reqwest::Url;
-use saucer::{anyhow, Context, Error};
-use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use std::{
-    collections::HashMap,
-    fmt::Debug,
-    io::{self, BufRead, BufReader, Write},
-    sync::mpsc::SyncSender,
-    time::{Duration, Instant},
-};
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[non_exhaustive]
-pub enum MessageKind {
-    AddSubgraph { subgraph_entry: SubgraphEntry },
-    UpdateSubgraph { subgraph_entry: SubgraphEntry },
-    RemoveSubgraph { subgraph_name: SubgraphName },
-    GetSubgraphs,
-    HealthCheck,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MessageSender {
-    ipc_socket_addr: String,
-    is_main_session: bool,
-}
-
-impl MessageSender {
-    pub fn new(ipc_socket_addr: &str, is_main_session: bool) -> Self {
-        Self {
-            ipc_socket_addr: ipc_socket_addr.to_string(),
-            is_main_session,
-        }
-    }
-
-    pub fn new_subgraph(ipc_socket_addr: &str) -> Self {
-        Self::new(ipc_socket_addr, false)
-    }
-
-    fn should_message(&self) -> bool {
-        !self.is_main_session
-    }
-
-    pub fn add_subgraph(&self, subgraph: &SubgraphDefinition) -> Result<()> {
-        if self.should_message() {
-            eprintln!(
-                "notifying main `rover dev` session about new subgraph '{}'",
-                &subgraph.name
-            );
-        }
-        self.socket_message::<()>(&MessageKind::AddSubgraph {
-            subgraph_entry: entry_from_definition(subgraph)?,
-        })?;
-        Ok(())
-    }
-
-    pub fn update_subgraph(&self, subgraph: &SubgraphDefinition) -> Result<()> {
-        if self.should_message() {
-            eprintln!(
-                "notifying main `rover dev` session about updated subgraph '{}'",
-                &subgraph.name
-            );
-        }
-        self.socket_message::<()>(&MessageKind::UpdateSubgraph {
-            subgraph_entry: entry_from_definition(subgraph)?,
-        })?;
-        Ok(())
-    }
-
-    pub fn remove_subgraph(&self, subgraph_name: &SubgraphName) -> Result<()> {
-        if self.should_message() {
-            eprintln!(
-                "notifying main `rover dev` session about removed subgraph '{}'",
-                &subgraph_name
-            );
-            self.socket_message::<()>(&MessageKind::RemoveSubgraph {
-                subgraph_name: subgraph_name.to_string(),
-            })?;
-        }
-
-        Ok(())
-    }
-
-    pub fn session_subgraphs(&self) -> Option<Vec<SubgraphKey>> {
-        if let Ok(Some(subgraphs)) =
-            self.socket_message::<Vec<SubgraphKey>>(&MessageKind::GetSubgraphs)
-        {
-            tracing::info!(
-                "the main `rover dev` session currently has {} subgraphs",
-                subgraphs.len()
-            );
-            Some(subgraphs)
-        } else {
-            tracing::info!("initializing the main `rover dev` session",);
-            None
-        }
-    }
-
-    pub fn health_check(&self) -> Result<()> {
-        loop {
-            if let Err(e) = self.socket_message::<()>(&MessageKind::HealthCheck) {
-                break Err(e);
-            }
-            std::thread::sleep(Duration::from_secs(1));
-        }
-    }
-
-    pub fn socket_message<T>(&self, message: &MessageKind) -> Result<Option<T>>
-    where
-        T: Serialize + DeserializeOwned + Debug,
-    {
-        match self.connect() {
-            Ok(stream) => {
-                stream
-                    .set_nonblocking(true)
-                    .context("could not set socket to non-blocking mode")?;
-                let mut stream = BufReader::new(stream);
-
-                // send our message over the socket
-                socket_write(message, &mut stream)?;
-
-                // wait for our message to be read by the other socket handler
-                // then read the response that was written back to the socket
-                socket_read(&mut stream)
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    fn connect(&self) -> Result<LocalSocketStream> {
-        LocalSocketStream::connect(&*self.ipc_socket_addr).map_err(|_| {
-            RoverError::new(anyhow!(
-                "the main `rover dev` session has been killed, shutting down"
-            ))
-        })
-    }
-
-    pub fn is_main_session(&self) -> bool {
-        self.is_main_session
-    }
-}
-
-pub type SubgraphName = String;
-pub type SubgraphUrl = Url;
-pub type SubgraphSdl = String;
-pub type SubgraphKey = (SubgraphName, SubgraphUrl);
-pub type SubgraphEntry = (SubgraphKey, SubgraphSdl);
-
-fn sdl_from_definition(subgraph_definition: &SubgraphDefinition) -> SubgraphSdl {
-    subgraph_definition.sdl.to_string()
-}
-
-fn name_from_definition(subgraph_definition: &SubgraphDefinition) -> SubgraphName {
-    subgraph_definition.name.to_string()
-}
-
-fn url_from_definition(subgraph_definition: &SubgraphDefinition) -> Result<SubgraphUrl> {
-    Ok(subgraph_definition.url.parse()?)
-}
-
-fn key_from_definition(subgraph_definition: &SubgraphDefinition) -> Result<SubgraphKey> {
-    Ok((
-        name_from_definition(subgraph_definition),
-        url_from_definition(subgraph_definition)?,
-    ))
-}
-
-fn entry_from_definition(subgraph_definition: &SubgraphDefinition) -> Result<SubgraphEntry> {
-    Ok((
-        key_from_definition(subgraph_definition)?,
-        sdl_from_definition(subgraph_definition),
-    ))
-}
-
-#[derive(Debug)]
-pub struct MessageReceiver {
-    subgraphs: HashMap<SubgraphKey, SubgraphSdl>,
-    socket_addr: String,
-    compose_runner: ComposeRunner,
-}
-
-impl MessageReceiver {
-    pub fn new(socket_addr: &str, compose_runner: ComposeRunner) -> Result<Self> {
-        if let Ok(stream) = LocalSocketStream::connect(socket_addr) {
-            // write to the socket so we don't make the other session deadlock waiting on a message
-            let mut stream = BufReader::new(stream);
-            let _ = socket_write(&(), &mut stream);
-            Err(RoverError::new(anyhow!(
-                "there is already a main `rover dev` session"
-            )))
-        } else {
-            Ok(Self {
-                subgraphs: HashMap::new(),
-                socket_addr: socket_addr.to_string(),
-                compose_runner,
-            })
-        }
-    }
-
-    pub fn supergraph_config(&self) -> SupergraphConfig {
-        let mut supergraph_config: SupergraphConfig = self
-            .subgraphs
-            .iter()
-            .map(|((name, url), sdl)| SubgraphDefinition::new(name, url.to_string(), sdl))
-            .collect::<Vec<SubgraphDefinition>>()
-            .into();
-        supergraph_config.set_federation_version(FederationVersion::LatestFedTwo);
-        supergraph_config
-    }
-
-    pub fn add_subgraph(&mut self, subgraph_entry: &SubgraphEntry) -> Result<()> {
-        let ((name, url), sdl) = subgraph_entry;
-        if self
-            .subgraphs
-            .get(&(name.to_string(), url.clone()))
-            .is_some()
-        {
-            Err(RoverError::new(anyhow!(
-                "subgraph with name '{}' and url '{}' already exists",
-                &name,
-                &url
-            )))
-        } else {
-            self.subgraphs
-                .insert((name.to_string(), url.clone()), sdl.to_string());
-            Ok(())
-        }
-    }
-
-    pub fn update_subgraph(&mut self, subgraph_entry: &SubgraphEntry) -> Result<()> {
-        let ((name, url), sdl) = subgraph_entry;
-        if let Some(prev_sdl) = self.subgraphs.get_mut(&(name.to_string(), url.clone())) {
-            if prev_sdl != sdl {
-                *prev_sdl = sdl.to_string();
-            }
-            Ok(())
-        } else {
-            self.add_subgraph(subgraph_entry)
-        }
-    }
-
-    pub fn remove_subgraph(&mut self, subgraph_name: &SubgraphName) -> Result<()> {
-        let mut found = None;
-        for (name, url) in self.subgraphs.keys() {
-            if name == subgraph_name {
-                found = Some((name, url));
-                break;
-            }
-        }
-        if let Some((name, url)) = found {
-            self.subgraphs.remove(&(name.to_string(), url.clone()));
-            Ok(())
-        } else {
-            Err(RoverError::new(anyhow!(
-                "subgraph with name '{}' does not exist",
-                &subgraph_name,
-            )))
-        }
-    }
-
-    pub fn receive_messages(
-        &mut self,
-        ready_sender: SyncSender<()>,
-        compose_sender: SyncSender<ComposeResult>,
-    ) -> Result<()> {
-        let listener = LocalSocketListener::bind(&*self.socket_addr).with_context(|| {
-            format!(
-                "could not start local socket server at {}",
-                &self.socket_addr
-            )
-        })?;
-        ready_sender.send(()).unwrap();
-        tracing::info!(
-            "connected to socket {}, waiting for messages",
-            &self.socket_addr
-        );
-        listener
-            .incoming()
-            .filter_map(handle_socket_error)
-            .for_each(|stream| {
-                let mut stream = BufReader::new(stream);
-                tracing::info!("received incoming socket connection");
-                let was_composed = self.compose_runner.has_composed();
-                match socket_read::<MessageKind>(&mut stream) {
-                    Ok(Some(message)) => match message {
-                        MessageKind::AddSubgraph { subgraph_entry } => {
-                            eprintln!("adding subgraph '{}'", &subgraph_entry.0 .0);
-                            let _ = self
-                                .add_subgraph(&subgraph_entry)
-                                .map(|_| {
-                                    let _ =
-                                        self.compose_runner.run(self).map_err(log_err_and_continue);
-                                })
-                                .map_err(log_err_and_continue);
-                            let _ = socket_write(&(), &mut stream).map_err(log_err_and_continue);
-                        }
-                        MessageKind::UpdateSubgraph { subgraph_entry } => {
-                            eprintln!("updating subgraph'{}'", &subgraph_entry.0 .0);
-                            let _ = self
-                                .update_subgraph(&subgraph_entry)
-                                .map(|_| {
-                                    let _ =
-                                        self.compose_runner.run(self).map_err(log_err_and_continue);
-                                })
-                                .map_err(log_err_and_continue);
-                            let _ = socket_write(&(), &mut stream).map_err(log_err_and_continue);
-                        }
-                        MessageKind::RemoveSubgraph { subgraph_name } => {
-                            eprintln!("removing subgraph'{}'", &subgraph_name);
-                            let _ = self.remove_subgraph(&subgraph_name).map(|_| {
-                                let _ = self.compose_runner.run(self).map_err(log_err_and_continue);
-                            });
-                            let _ = socket_write(&(), &mut stream).map_err(log_err_and_continue);
-                        }
-                        MessageKind::GetSubgraphs => {
-                            eprintln!("notifying new `rover dev` session about existing subgraphs");
-                            let _ = socket_write(&self.get_subgraphs(), &mut stream)
-                                .map_err(log_err_and_continue);
-                        }
-                        MessageKind::HealthCheck => (),
-                    },
-                    Ok(None) => {}
-                    Err(e) => log_err_and_continue(e),
-                }
-
-                let has_composed = self.compose_runner.has_composed();
-
-                if has_composed && !was_composed {
-                    compose_sender.send(ComposeResult::Succeed).unwrap();
-                } else if !has_composed && was_composed {
-                    compose_sender.send(ComposeResult::Fail).unwrap();
-                }
-            });
-        Ok(())
-    }
-
-    pub fn get_subgraphs(&self) -> Vec<SubgraphKey> {
-        self.subgraphs.keys().cloned().collect()
-    }
-}
-
-pub enum ComposeResult {
-    Succeed,
-    Fail,
-    Kill,
-}
-
-fn handle_socket_error(conn: io::Result<LocalSocketStream>) -> Option<LocalSocketStream> {
-    match conn {
-        Ok(val) => Some(val),
-        Err(error) => {
-            eprintln!("incoming connection failed: {}", error);
-            None
-        }
-    }
-}
-
-pub(crate) fn socket_read<B>(stream: &mut BufReader<LocalSocketStream>) -> Result<Option<B>>
-where
-    B: Serialize + DeserializeOwned + Debug,
-{
-    tracing::info!("\n----    RECEIVE     ----\n");
-    let mut incoming_message = String::new();
-
-    let now = Instant::now();
-
-    let result = loop {
-        if now.elapsed() > Duration::from_secs(5) {
-            break Err(anyhow!(
-                "could not read incoming message from the main `rover dev` session after 5 seconds"
-            ));
-        }
-
-        match stream.read_line(&mut incoming_message) {
-            Ok(_) => {
-                break Ok(
-                    if incoming_message.is_empty() || &incoming_message == "null\n" {
-                        None
-                    } else {
-                        let incoming_message: B = serde_json::from_str(&incoming_message)
-                            .with_context(|| {
-                                format!(
-                                    "incoming message '{}' was not valid JSON",
-                                    &incoming_message
-                                )
-                            })?;
-                        tracing::info!("\n{:?}\n", &incoming_message);
-                        Some(incoming_message)
-                    },
-                )
-            }
-            Err(e) => {
-                if !matches!(e.kind(), io::ErrorKind::WouldBlock) {
-                    break Err(Error::new(e).context("could not read incoming message"));
-                }
-            }
-        }
-
-        std::thread::sleep(Duration::from_millis(500));
-    }?;
-
-    tracing::info!("\n====   END RECEIVE    ====\n");
-    Ok(result)
-}
-
-pub(crate) fn socket_write<A>(message: &A, stream: &mut BufReader<LocalSocketStream>) -> Result<()>
-where
-    A: Serialize + DeserializeOwned + Debug,
-{
-    tracing::info!("\n----      SEND      ----\n");
-    tracing::info!("\n{:?}\n", &message);
-    let outgoing_json = serde_json::to_string(message)
-        .with_context(|| format!("could not convert outgoing message {:?} to json", &message))?;
-    let outgoing_string = format!("{}\n", &outgoing_json);
-    stream
-        .get_mut()
-        .write_all(outgoing_string.as_bytes())
-        .context("could not write outgoing message to socket")?;
-    tracing::info!("\n====    END SEND     ====\n");
-    Ok(())
-}
